backend/djanql/requirements.txt
ariadne==0.13.0
asgiref==3.4.1
cached-property==1.5.2
certifi==2021.5.30
click==8.0.1
Django==3.1.13
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-cors-headers==3.8.0
django-cursor-pagination==0.1.5
django-filer==2.0.2
django-formtools==2.3
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-tinymce==3.3.0
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.7.1
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
graphql-core==3.1.6
h11==0.12.0
html5lib==1.1
importlib-metadata==4.8.1
Pillow==8.3.1
psycopg2-binary==2.9.1
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
starlette==0.14.2
strawberry-graphql==0.78.1
typing-extensions==3.10.0.0
Unidecode==1.1.2
urllib3==1.26.6
uvicorn==0.15.0
webencodings==0.5.1
zipp==3.5.0
